requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
yt-dlp==2025.9.26
requests==2.31.0
python-dotenv==1.0.0
google-api-python-client==2.108.0